Use a first-class AdvisorySource
Establish the convention of having `jlsec_` fields for database specific ones. Store the sources in a vector that represents the sequence of events from which the JLSEC advisory was cobbled together. Use an optional `fields` to represent only updating some subset of fields. This should help with round-trippability, fix the issues with string/DateTime timestamps, and hopefully make it easier to avoid clobbering our updates if the upstream source has not changed since our last import.